console: Split gateway-connection view and logic
